requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
